refactor so uframe routes stay in uframe controller, and helper data modifiers stay in uframe data
