+++ ooiservices/app/uframe/data.py
@@ -13,6 +13,8 @@ import time
 from dateutil.parser import parse
 from datetime import datetime
 
+from ooiservices.app import cache
+
 def gen_data(start_date, end_date, sampling_rate, mean, std_dev):
     '''
     Returns a dictionary that contains the x coordinate time and the y
@@ -35,3 +37,180 @@ def gen_data(start_date, end_date, sampling_rate, mean, std_dev):
 
     return {'size' : x.shape[0], 'start_time':iso0, 'end_time':iso1, 'cols':['x','y'], 'rows':row_order_xy.tolist()}
 
+@cache.memoize(timeout=3600)
+def plot_time_series(fig, ax, x, y, fill=False, title='', ylabel='',
+                         title_font={}, axis_font={}, **kwargs):
+
+    if not title_font:
+        title_font = title_font_default
+    if not axis_font:
+        axis_font = axis_font_default
+
+    h = ppl.plot(ax, x, y, **kwargs)
+    ppl.scatter(ax, x, y, **kwargs)
+    get_time_label(ax, x)
+    fig.autofmt_xdate()
+
+    if ylabel:
+        ax.set_ylabel(ylabel, **axis_font)
+    if title:
+        ax.set_title(title, **title_font)
+    if 'degree' in ylabel:
+        ax.set_ylim([0, 360])
+    ax.grid(True)
+    if fill:
+        miny = min(ax.get_ylim())
+        ax.fill_between(x, y, miny+1e-7, facecolor = h[0].get_color(), alpha=0.15)
+    # plt.subplots_adjust(top=0.85)
+    plt.tight_layout()
+
+def get_time_label(ax, dates):
+    '''
+    Custom date axis formatting
+    '''
+    def format_func(x, pos=None):
+        x = mdates.num2date(x)
+        if pos == 0:
+            fmt = '%Y-%m-%d %H:%M'
+        else:
+            fmt = '%H:%M'
+        label = x.strftime(fmt)
+        # label = label.rstrip("0")
+        # label = label.rstrip(".")
+        return label
+    day_delta = (max(dates)-min(dates)).days
+
+    if day_delta < 1:
+        ax.xaxis.set_major_formatter(FuncFormatter(format_func))
+    else:
+        # pass
+        major = mdates.AutoDateLocator()
+        formt = mdates.AutoDateFormatter(major, defaultfmt=u'%Y-%m-%d')
+        formt.scaled[1.0] = '%Y-%m-%d'
+        formt.scaled[30] = '%Y-%m'
+        formt.scaled[1./24.] = '%Y-%m-%d %H:%M:%S'
+        # formt.scaled[1./(24.*60.)] = FuncFormatter(format_func)
+        ax.xaxis.set_major_locator(major)
+        ax.xaxis.set_major_formatter(formt)
+
+def plot_scatter(fig, ax, x, y, title='', xlabel='', ylabel='',
+                     title_font={}, axis_font={}, **kwargs):
+
+        if not title_font:
+            title_font = title_font_default
+        if not axis_font:
+            axis_font = axis_font_default
+
+        ppl.scatter(ax, x, y, **kwargs)
+        if xlabel:
+            ax.set_xlabel(xlabel, labelpad=10, **axis_font)
+        if ylabel:
+            ax.set_ylabel(ylabel, labelpad=10, **axis_font)
+        ax.set_title(title, **title_font)
+        ax.grid(True)
+        ax.set_aspect(1./ax.get_data_ratio())  # make axes square
+
+def _get_data_type(data_input):
+    '''
+    gets the data type in a format google charts understands
+    '''
+    if data_input is float or data_input is int:
+        return "number"
+    elif data_input is str or data_input is unicode:
+        return "string"
+    else:
+        return "unknown"
+
+def _get_annotation(instrument_name, stream_name):
+    annotations = Annotation.query.filter_by(instrument_name=instrument_name, stream_name=stream_name).all()
+    return [annotation.to_json() for annotation in annotations]
+
+def _get_col_outline(data,pref_timestamp,inital_fields,requested_field):
+    '''
+    gets the column outline for the google chart response, figures out what annotations are required where...
+    '''
+    data_fields = []
+    data_field_list= []
+    #used to cound the fields, used for annotations
+    field_count = 1
+    #loop and generate inital col dict
+    for field in inital_fields:
+        if field == pref_timestamp:
+            d_type = "datetime"
+        elif field in FIELDS_IGNORE or str(field).endswith('_timestamp'):
+            continue
+        else:
+            if requested_field is not None:
+                if field == requested_field:
+                    d_type = _get_data_type(type(data[0][field]))
+                else:
+                    continue
+            else:
+                #map the data types to the correct data type for google charts
+                d_type = _get_data_type(type(data[0][field]))
+
+        data_field_list.append(field)
+        data_fields.append({"id": "",
+                            "label": field,
+                            "type":  d_type})
+
+    return data_fields,data_field_list
+
+def _get_annotation_content(annotation_field, pref_timestamp, annotations_list, d, data_field):
+    '''
+    creates the annotation content for a given field
+    '''
+    #right now x and y are timeseries data
+    for an in annotations_list:
+        if an['field_x'] == pref_timestamp or an['field_y'] == data_field:
+            # and and y value
+            an_date_time = datetime.datetime.strptime(an['pos_x'], "%Y-%m-%dT%H:%M:%S")
+            an_int_date_time = int(an_date_time.strftime("%s"))
+
+            if int(d['fixed_dt']) == an_int_date_time:
+                if annotation_field == "annotation":
+                    return {"v":an["title"]}
+                elif annotation_field == "annotationText":
+                    return {"v":an['comment']}
+
+    #return nothing
+    return {"v":None,"f":None}
+
+def make_cache_key():
+    return urlencode(request.args)
+
+@cache.memoize(timeout=3600)
+def get_uframe_streams():
+    '''
+    Lists all the streams
+    '''
+    try:
+        UFRAME_DATA = current_app.config['UFRAME_URL'] + '/sensor/m2m/inv'
+        response = requests.get(UFRAME_DATA)
+        return response
+    except:
+        return internal_server_error('uframe connection cannot be made.')
+
+@cache.memoize(timeout=3600)
+def get_uframe_stream(stream):
+    '''
+    Lists the reference designators for the streams
+    '''
+    try:
+        UFRAME_DATA = current_app.config['UFRAME_URL'] + '/sensor/m2m/inv'
+        response = requests.get("/".join([UFRAME_DATA,stream]))
+        return response
+    except:
+        return internal_server_error('uframe connection cannot be made.')
+
+@cache.memoize(timeout=3600)
+def get_uframe_stream_contents(stream, ref):
+    '''
+    Gets the stream contents
+    '''
+    try:
+        UFRAME_DATA = current_app.config['UFRAME_URL'] + '/sensor/m2m/inv'
+        response =  requests.get("/".join([UFRAME_DATA,stream,ref]))
+        return response
+    except:
+        return internal_server_error('uframe connection cannot be made.')
